+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/login/elder/screen/LoginElderInfoScreen.kt
@@ -1,4 +1,4 @@
-package com.konkuk.medicarecall.ui.feature.login.senior.screen
+package com.konkuk.medicarecall.ui.feature.login.elder.screen
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -44,10 +44,10 @@ import com.konkuk.medicarecall.R
 import com.konkuk.medicarecall.ui.common.component.CTAButton
 import com.konkuk.medicarecall.ui.common.component.DefaultSnackBar
 import com.konkuk.medicarecall.ui.common.util.isValidDate
-import com.konkuk.medicarecall.ui.feature.login.info.component.LoginBackButton
-import com.konkuk.medicarecall.ui.feature.login.senior.component.ElderChip
-import com.konkuk.medicarecall.ui.feature.login.senior.component.ElderInputForm
-import com.konkuk.medicarecall.ui.feature.login.senior.viewmodel.LoginElderViewModel
+import com.konkuk.medicarecall.ui.feature.login.myinfo.component.LoginBackButton
+import com.konkuk.medicarecall.ui.feature.login.elder.component.ElderChip
+import com.konkuk.medicarecall.ui.feature.login.elder.component.ElderInputForm
+import com.konkuk.medicarecall.ui.feature.login.elder.viewmodel.LoginElderViewModel
 import com.konkuk.medicarecall.ui.theme.MediCareCallTheme
 import com.konkuk.medicarecall.ui.type.CTAButtonType
 import kotlinx.coroutines.delay
