+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/login/elder/screen/LoginElderMedInfoScreen.kt
@@ -1,4 +1,4 @@
-package com.konkuk.medicarecall.ui.feature.login.senior.screen
+package com.konkuk.medicarecall.ui.feature.login.elder.screen
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -37,8 +37,8 @@ import com.konkuk.medicarecall.ui.common.component.DefaultDropdown
 import com.konkuk.medicarecall.ui.common.component.DefaultSnackBar
 import com.konkuk.medicarecall.ui.common.component.DiseaseNamesItem
 import com.konkuk.medicarecall.ui.common.component.MedicationItem
-import com.konkuk.medicarecall.ui.feature.login.info.component.LoginBackButton
-import com.konkuk.medicarecall.ui.feature.login.senior.viewmodel.LoginElderViewModel
+import com.konkuk.medicarecall.ui.feature.login.myinfo.component.LoginBackButton
+import com.konkuk.medicarecall.ui.feature.login.elder.viewmodel.LoginElderViewModel
 import com.konkuk.medicarecall.ui.theme.MediCareCallTheme
 import com.konkuk.medicarecall.ui.type.CTAButtonType
 import com.konkuk.medicarecall.ui.type.HealthIssueType
