+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/login/elder/viewmodel/LoginElderUiState.kt
@@ -1,4 +1,4 @@
-package com.konkuk.medicarecall.ui.feature.login.senior.viewmodel
+package com.konkuk.medicarecall.ui.feature.login.elder.viewmodel
 
 import com.konkuk.medicarecall.ui.model.ElderData
