+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/login/myinfo/screen/LoginPhoneScreen.kt
@@ -1,4 +1,4 @@
-package com.konkuk.medicarecall.ui.feature.login.info.screen
+package com.konkuk.medicarecall.ui.feature.login.myinfo.screen
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -31,8 +31,8 @@ import com.konkuk.medicarecall.ui.common.component.CTAButton
 import com.konkuk.medicarecall.ui.common.component.DefaultSnackBar
 import com.konkuk.medicarecall.ui.common.component.DefaultTextField
 import com.konkuk.medicarecall.ui.common.util.PhoneNumberVisualTransformation
-import com.konkuk.medicarecall.ui.feature.login.info.component.LoginBackButton
-import com.konkuk.medicarecall.ui.feature.login.info.viewmodel.LoginInfoViewModel
+import com.konkuk.medicarecall.ui.feature.login.myinfo.component.LoginBackButton
+import com.konkuk.medicarecall.ui.feature.login.myinfo.viewmodel.LoginInfoViewModel
 import com.konkuk.medicarecall.ui.theme.MediCareCallTheme
 import com.konkuk.medicarecall.ui.type.CTAButtonType
 import kotlinx.coroutines.launch
