+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/login/myinfo/screen/LoginStartScreen.kt
@@ -1,4 +1,4 @@
-package com.konkuk.medicarecall.ui.feature.login.info.screen
+package com.konkuk.medicarecall.ui.feature.login.myinfo.screen
 
 import android.annotation.SuppressLint
 import android.app.Activity
@@ -28,7 +28,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.konkuk.medicarecall.R
 import org.koin.androidx.compose.koinViewModel
 import com.konkuk.medicarecall.ui.common.component.CTAButton
-import com.konkuk.medicarecall.ui.feature.login.info.viewmodel.LoginInfoViewModel
+import com.konkuk.medicarecall.ui.feature.login.myinfo.viewmodel.LoginInfoViewModel
 import com.konkuk.medicarecall.ui.model.NavigationDestination
 import com.konkuk.medicarecall.ui.theme.MediCareCallTheme
 import com.konkuk.medicarecall.ui.type.CTAButtonType
