+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/login/navigation/LoginNavigation.kt
@@ -7,15 +7,15 @@ import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.konkuk.medicarecall.ui.common.extension.sharedViewModel
 import com.konkuk.medicarecall.ui.feature.login.calltime.screen.CallTimeScreen
-import com.konkuk.medicarecall.ui.feature.login.info.screen.LoginMyInfoScreen
-import com.konkuk.medicarecall.ui.feature.login.info.screen.LoginPhoneScreen
-import com.konkuk.medicarecall.ui.feature.login.info.screen.LoginStartScreen
-import com.konkuk.medicarecall.ui.feature.login.info.screen.LoginVerificationScreen
-import com.konkuk.medicarecall.ui.feature.login.info.viewmodel.LoginInfoViewModel
+import com.konkuk.medicarecall.ui.feature.login.myinfo.screen.LoginMyInfoScreen
+import com.konkuk.medicarecall.ui.feature.login.myinfo.screen.LoginPhoneScreen
+import com.konkuk.medicarecall.ui.feature.login.myinfo.screen.LoginStartScreen
+import com.konkuk.medicarecall.ui.feature.login.myinfo.screen.LoginVerificationScreen
+import com.konkuk.medicarecall.ui.feature.login.myinfo.viewmodel.LoginInfoViewModel
 import com.konkuk.medicarecall.ui.feature.login.payment.screen.LoginFinishScreen
-import com.konkuk.medicarecall.ui.feature.login.senior.screen.LoginElderMedInfoScreen
-import com.konkuk.medicarecall.ui.feature.login.senior.screen.LoginElderScreen
-import com.konkuk.medicarecall.ui.feature.login.senior.viewmodel.LoginElderViewModel
+import com.konkuk.medicarecall.ui.feature.login.elder.screen.LoginElderMedInfoScreen
+import com.konkuk.medicarecall.ui.feature.login.elder.screen.LoginElderScreen
+import com.konkuk.medicarecall.ui.feature.login.elder.viewmodel.LoginElderViewModel
 import com.konkuk.medicarecall.ui.navigation.Route
 
 fun NavController.navigateToLoginStart() {
